leverage the states and the functions
