+++ react/2-react-fundamental-1/src/pages/Main.jsx
@@ -1,121 +1,55 @@
-import { Component } from 'react'
+import PropTypes from 'prop-types'
 
 // COMPONENTS
 import CreateNoteItem from 'components/CreateNoteItem'
-import Header from 'components/Header'
 import NoteList from 'components/NoteList'
 
 // MUIS
 import Stack from '@mui/material/Stack'
 
-// UTILITIES
-import { getInitialData } from 'utilities/data'
+const Main = (props) => {
+  const { 
+    filteredNoteList,
+    onAddNewNote,
+    onDeleteNote,
+    onArchiveNote,
+  } = props
 
-class Main extends Component {
-  constructor(props) {
-    super(props)
+  return (
+    <>        
+      {/* CREATE A NOTE ITEM */}
+      <CreateNoteItem onSubmitButtonClick={onAddNewNote}/>
 
-    this.state = {
-      search: '',
-      noteList: getInitialData(),
-      filteredNoteList: getInitialData(),
-    }
+      <Stack direction='row'>
+        {/* ACTIVE NOTES */}
+        <NoteList 
+          noteList={filteredNoteList.filter(item => !item.archived)}
+          type='active'
+          onDeleteButtonClick={onDeleteNote}
+          onArchiveButtonClick={onArchiveNote}
+        />
 
-    this.onAddNewNoteHandler = this.onAddNewNoteHandler.bind(this)
-    this.onDeleteNoteHandler = this.onDeleteNoteHandler.bind(this)
-    this.onArchiveNoteHandler = this.onArchiveNoteHandler.bind(this)
-    this.onSearchChangeHandler = this.onSearchChangeHandler.bind(this)
-  }
+        {/* ARCHIVED NOTES */}
+        <NoteList 
+          noteList={filteredNoteList.filter(item => item.archived)}
+          type='archived'
+          onDeleteButtonClick={onDeleteNote}
+          onArchiveButtonClick={onArchiveNote}
+        />
+      </Stack>
+    </>
+  )
+}
 
-  onAddNewNoteHandler ({ title, body }) {
-    this.setState((prevState) => {
-      const newNoteList = [
-        ...prevState.noteList,
-        {
-          id: +new Date(),
-          title,
-          body,
-          archived: false,
-          createdAt: new Date(),
-        },
-      ]
+Main.defaultProps = {
+  filteredNoteList: [],
+}
 
-      return {
-        ...prevState,
-        noteList: newNoteList,
-        filteredNoteList: newNoteList.filter(item => item.title.toLowerCase().includes(this.state.search.toLowerCase())),
-      }
-    })
-  }
-
-  onDeleteNoteHandler (id) {
-    this.setState((prevState) => {
-      const newNoteList = prevState.noteList.filter(item => item.id !== id)
-
-      return {
-        ...prevState,
-        noteList: newNoteList,
-        filteredNoteList: newNoteList.filter(item => item.title.toLowerCase().includes(this.state.search.toLowerCase())),
-      }
-    })
-  }
-
-  onArchiveNoteHandler (id) {
-    this.setState((prevState) => {
-      const newNoteList = prevState.noteList.map(item => {
-        return {
-          ...item,
-          archived: item.id === id ? !item.archived : item.archived,
-        }
-      })
-
-      return {
-        ...prevState,
-        noteList: newNoteList,
-        filteredNoteList: newNoteList.filter(item => item.title.toLowerCase().includes(this.state.search.toLowerCase())),
-      }
-    })
-  }
-
-  onSearchChangeHandler (search) {
-    this.setState((prevState) => {
-      return {
-        ...prevState,
-        search,
-        filteredNoteList: prevState.noteList.filter(item => item.title.toLowerCase().includes(search.toLowerCase())),
-      }
-    })
-  }
-
-  render() {
-    return (
-      <>
-        {/* HEADER */}
-        <Header onSearchChange={this.onSearchChangeHandler}/>
-        
-        {/* CREATE A NOTE ITEM */}
-        <CreateNoteItem onSubmitButtonClick={this.onAddNewNoteHandler}/>
-
-        <Stack direction='row'>
-          {/* ACTIVE NOTES */}
-          <NoteList 
-            noteList={this.state.filteredNoteList.filter(item => !item.archived)}
-            type='active'
-            onDeleteButtonClick={this.onDeleteNoteHandler}
-            onArchiveButtonClick={this.onArchiveNoteHandler}
-          />
-
-          {/* ARCHIVED NOTES */}
-          <NoteList 
-            noteList={this.state.filteredNoteList.filter(item => item.archived)}
-            type='archived'
-            onDeleteButtonClick={this.onDeleteNoteHandler}
-            onArchiveButtonClick={this.onArchiveNoteHandler}
-          />
-        </Stack>
-      </>
-    )
-  }
+Main.propTypes = {
+  filteredNoteList: PropTypes.array.isRequired,
+  onAddNewNote: PropTypes.func.isRequired,
+  onDeleteNote: PropTypes.func.isRequired,
+  onArchiveNote: PropTypes.func.isRequired,
 }
 
 export default Main
